+++ c/options.h
@@ -12,25 +12,31 @@
 
 namespace config {
 
-int get_nthreads();
-void set_nthreads(int nth);
-
 extern PyObject* logger;
-PyObject* get_core_logger();
+extern int nthreads;
+extern size_t sort_insert_method_threshold;
+extern size_t sort_thread_multiplier;
+extern size_t sort_max_chunk_length;
+extern int8_t sort_max_radix_bits;
+extern int8_t sort_over_radix_bits;
+
+void set_nthreads(int nth);
 void set_core_logger(PyObject*);
+void set_sort_insert_method_threshold(int64_t n);
+void set_sort_thread_multiplier(int64_t n);
+void set_sort_max_chunk_length(int64_t n);
+void set_sort_max_radix_bits(int64_t n);
+void set_sort_over_radix_bits(int64_t n);
+
+
+DECLARE_FUNCTION(
+  set_option,
+  "set_option(name, value)\n\n"
+  "Set core option `name` to the given `value`. The `name` must be a string,\n"
+  "and it must be one of the recognizable option names. If not, an exception\n"
+  "will be raised. The allowed `value`s depend on the option being set.",
+  dt_OPTIONS_cc)
+
 };
-
-
-
-DECLARE_FUNCTION(
-  set_nthreads,
-  "set_nthreads(nth)\n\n",
-  dt_OPTIONS_cc)
-
-DECLARE_FUNCTION(
-  set_core_logger,
-  "set_core_logger(logger)\n\n",
-  dt_OPTIONS_cc)
-
 
 #endif